Foundation, Modeling - NCollection modernization and BRepGraph overhaul (#1212)
- Migrate NCollection map/sequence/array size APIs from int to size_t; Size() returns size_t, Length() remains the int accessor; int overloads delegate through NbBucketsFromInt() with a negative-input guard.
- Add NCollection_LinearVector: contiguous flat-buffer dynamic array with Standard::Reallocate-based growth for trivial types and move-construction for non-trivial types.
- Rewrite NCollection_DynamicArray on top of LinearVector<T*>; switch to power-of-two block sizing with precomputed shift/mask.
- Rework NCollection_BaseMap iterator with a forward findFirst() helper (no negative-bucket arithmetic); unify ReSize/BeginResize/EndResize on size_t.
- Enable thread-safe fast path in NCollection_IncAllocator via std::shared_mutex + CAS bump allocation on atomic AvailableSize; exclusive lock is taken only for new-block allocation and list reordering.
- Remove NCollection_BasePointerVector (superseded by NCollection_LinearVector).
- Remove NCollection_BaseMap::Statistics (unused).
- Add <cstddef> include in NCollection_Primes.hxx so size_t resolves on clean builds.
- Document iterator invalidation contract on NCollection_LinearVector.
- Unify BRepGraph programmatic mutation behind EditorView: delete BuilderView (2648+552 lines); EditorView (3186+930 lines) and EditorView_Mut.cxx own both structural creation (Add*/Remove*) and field-level RAII-scoped mutation (Mut*()) with automatic OwnGen and SubtreeGen propagation.
- Add BRepGraph_MeshCache and BRepGraph_MeshView: two-tier mesh storage separating algorithm-derived caches from persistent (definition) triangulations; freshness is keyed on FaceDef.OwnGen; cache writes do not mutate the model.
- Document the MeshCache invalidation contract in BRepGraph_MeshCache.hxx (which mutations bump Face.OwnGen and how markRepModified closes the loop for Surface/Triangulation reps).
- Add BRepGraph_RefsIterator (generic flat ref scan with RefTraits dispatch) and BRepGraph_ReverseIterator (typed parent-traversal wrappers over reverse-index vectors).
- Rework RefId entity model: replace inline refs with typed RefId vectors; add OccurrenceRef and Kind::Occurrence=7; BRepGraphInc_WireExplorer now requires a VertexRefLookup.
- Rename layers for consistency: BRepGraph_ParamLayer to BRepGraph_LayerParam, BRepGraph_RegularityLayer to BRepGraph_LayerRegularity; rename BRepGraphInc_Usage to BRepGraphInc_Instance.
- Replace RootNodeIds() with RootProductIds() returning product roots only.
- Update BRepGraph and BRepGraphInc READMEs; fix stale RootNodeIds reference.
- Sweep Size() to Length() renames across ~200 callers in TKG2d, TKG3d, TKMath, TKMesh, TKBO, TKOffset, TKShHealing, TKTopAlgo, TKBRep, TKService, TKV3d, TKOpenGl, TKMeshVS, TKDE*, TKXCAF, TKXSBase, TKLCAF, TKStd, and Draw harness.
- Add GTest coverage for new containers and the BRepGraph overhaul: NCollection_DynamicArray_Test, NCollection_LinearVector_Test, BRepGraph_Fuzz_Test, BRepGraph_Iterator_Test, BRepGraph_LayerIterator_Test, BRepGraph_MeshCache_Test, BRepGraph_MutGuard_Test, BRepGraph_ReplaceVertex_Test, BRepGraph_ReverseIterator_Test, BRepGraph_ScenarioMatrix_Test, BRepGraph_TypedIdDispatch_Test, BRepGraph_WireExplorer_Test.

+++ src/FoundationClasses/TKernel/NCollection/NCollection_FlatDataMap.hxx
@@ -223,7 +223,7 @@ public:
 
   //! Constructor with initial capacity hint
   //! @param theNbBuckets initial capacity (will be rounded up to power of 2)
-  explicit NCollection_FlatDataMap(const int theNbBuckets)
+  explicit NCollection_FlatDataMap(const size_t theNbBuckets)
       : mySlots(nullptr),
         myCapacity(0),
         mySize(0)
@@ -237,7 +237,7 @@ public:
   //! Constructor with custom hasher (copy).
   //! @param theHasher custom hasher instance
   //! @param theNbBuckets initial capacity hint
-  explicit NCollection_FlatDataMap(const Hasher& theHasher, const int theNbBuckets = 0)
+  explicit NCollection_FlatDataMap(const Hasher& theHasher, const size_t theNbBuckets = 0)
       : mySlots(nullptr),
         myCapacity(0),
         mySize(0),
@@ -252,7 +252,7 @@ public:
   //! Constructor with custom hasher (move).
   //! @param theHasher custom hasher instance (moved)
   //! @param theNbBuckets initial capacity hint
-  explicit NCollection_FlatDataMap(Hasher&& theHasher, const int theNbBuckets = 0)
+  explicit NCollection_FlatDataMap(Hasher&& theHasher, const size_t theNbBuckets = 0)
       : mySlots(nullptr),
         myCapacity(0),
         mySize(0),
@@ -363,10 +363,10 @@ public:
 public:
   // **************** Query methods ****************
 
-  //! Returns number of elements
-  int Size() const noexcept { return static_cast<int>(mySize); }
+  //! Returns number of elements.
+  size_t Size() const noexcept { return mySize; }
 
-  //! Returns number of elements
+  //! Returns number of elements (legacy int-returning API, convention shared with BaseMap).
   int Extent() const noexcept { return static_cast<int>(mySize); }
 
   //! Returns true if map is empty
